Rename kintoupdater into kinto_updater.
On PyPI it will be kinto-updater and it will be the kinto_updater as a python
module.
