+++ kinto_updater/signing.py
